% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bplot.r
\name{bplot}
\alias{bplot}
\title{bplot rolling average plotting function}
\usage{
bplot(t, x, y, ys = 1, ystep = 0, ylab = "", axis3 = "NA", w = 400,
  s = 80, f = "mean", scale = "linear", leg = "topleft", col = 1:100,
  max = "default", min = "default", feature = "data points")
}
\arguments{
\item{t}{Table of data.}

\item{x}{Variable to be binned on x-axis.}

\item{y}{Vector of variables to plot on y-axis.}

\item{ys}{Scale the y-axis variables first. Defaults to 1.}

\item{ystep}{Division of steps printed on y-axis scale. Defaults to 0.2.}

\item{ylab}{Label for Y-axis. Defaults to blank.}

\item{axis3}{Column of length variable to be plotted on 3rd axis. Defaults to "NA" (not shown).}

\item{w}{Window size (number of values per bin). Defaults to 400.}

\item{s}{Step size (number of values to move along for each bin). Defaults to 80.}

\item{f}{Function to apply to y values for each bin ("mean","median","sum"). Defaults to "mean".}

\item{scale}{Scale the x-axis ("linear"=linear spacing of points based on mean vaue of bin or "bins"=equal spacing between all bins). Defaults to "linear".}

\item{leg}{Postion of legend. Defaults to "topleft".}

\item{col}{Vector of colours to use in plots. Defaults to 1:100.}

\item{max, min}{Minimum and Maximum range of y-axis.}

\item{feature}{Name of features being binned. Defaults to "data points".}
}
\description{
This function orders and bins a variable in to groups defined by a window 
and step size then computes rolling averages for multipe variables across each bin.
}
\examples{
bplot()
}
\keyword{average}
\keyword{bin}
\keyword{plots.}

